requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.11.7
python-multipart==0.0.20
bcrypt==5.0.0
PyJWT==2.10.1
openpyxl==3.1.5
httpx==0.28.1
pytest==8.4.1
websockets==15.0.1
